Add network_security_config to allow proxy tools, code improvements (#12)
